cxc_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg==3.1.18
python-dotenv==1.0.1
